fix i64.popcnt

+++ module/module.go
@@ -79,7 +79,7 @@ func I64clz(x uint64) int32    { return int32(bits.LeadingZeros64(x)) }
 func I32ctz(x uint32) int32    { return int32(bits.TrailingZeros32(x)) }
 func I64ctz(x uint64) int32    { return int32(bits.TrailingZeros64(x)) }
 func I32popcnt(x uint32) int32 { return int32(bits.OnesCount32(x)) }
-func I64popcnt(x uint64) int32 { return int32(bits.OnesCount64(x)) }
+func I64popcnt(x uint64) int64 { return int64(bits.OnesCount64(x)) }
 
 // f64 ops
 func F64abs(x float64) float64         { return math.Abs(x) }
